Refactor taskBatcher.py module to lib/
