+++ lib/__init__.py
@@ -1,7 +1,32 @@
+"""
+EC codebase Python library (AKA the "frontend")
+
+Module mapping details:
+
+TODO: remove module mapping code when backwards-compatibility is no longer required.
+
+The below module mapping is required for backwards-compatibility with old pickle files
+generated from before the EC codebase refactor.
+
+The mapping uses the following pattern:
+
+    sys.modules[<old module path>] = <new module reference>
+
+This is because the previous structure of the codebase was completely flat, and when refactoring
+to a hierarchical files, loading previous pickle files no longer works properly. It is important
+to retain the ability to read old pickle files generated from official experiments. As a workaround,
+the old module paths are included below. A preferable alternative would be to export program state
+into JSON files instead of pickle files to avoid issues where the underlying classes change, so that
+could be a future improvement to this project. Until then, we use the module mapping workaround.
+
+For more info, see this StackOverflow answer: https://stackoverflow.com/a/2121918/2573242
+"""
 import sys
 
 from lib import ec
+from lib import primitiveGraph
 from lib import task
+from lib import taskBatcher
 from lib.domains.misc import algolispPrimitives, deepcoderPrimitives
 from lib.domains.misc import RobustFillPrimitives
 from lib.domains.misc import napsPrimitives
@@ -17,16 +42,9 @@ from lib.domains.arithmetic import arithmeticPrimitives
 from lib.domains.text import textPrimitives
 from lib.domains.text import makeTextTasks
 
-# TODO: commented out due to circular import error
-# from lib.domains.puddleworld import makePuddleworldTasks
-# from lib.domains.puddleworld import puddleworldPrimitives
-
-from lib import primitiveGraph
-
-# Required for backwards-compatibility with old pickle files
-# from before the EC codebase refactor:
 sys.modules['ec'] = ec
 sys.modules['task'] = task
+sys.modules['taskBatcher'] = taskBatcher
 sys.modules['algolispPrimitives'] = algolispPrimitives
 sys.modules['RobustFillPrimitives'] = RobustFillPrimitives
 sys.modules['napsPrimitives'] = napsPrimitives
@@ -42,6 +60,4 @@ sys.modules['makeListTasks'] = makeListTasks
 sys.modules['arithmeticPrimitives'] = arithmeticPrimitives
 sys.modules['textPrimitives'] = textPrimitives
 sys.modules['makeTextTasks'] = makeTextTasks
-# sys.modules['makePuddleworldTasks'] = makePuddleworldTasks
-# sys.modules['puddleworldPrimitives'] = puddleworldPrimitives
 sys.modules['primitiveGraph'] = primitiveGraph
